requirements.txt
anthropic
pdfplumber
weasyprint
google-generativeai
streamlit
python-dotenv